refactor: 상품 기능 패키지 이동 및 관련 로직 수정
